+++ feature_browse/src/main/java/ru/kggm/feature_browse/data/network/services/CharacterService.kt
@@ -15,6 +15,6 @@ interface CharacterService {
 
     @GET("character/{id}")
     fun getById(
-        @Path("id") id: Int
+        @Path("id") id: Long
     ): CompletableFuture<CharacterDto>
 }
