+++ feature_browse/src/main/java/ru/kggm/feature_browse/domain/repositories/CharacterRepository.kt
@@ -1,9 +1,9 @@
 package ru.kggm.feature_browse.domain.repositories
 
 import ru.kggm.feature_browse.domain.entities.CharacterEntity
-import ru.kggm.feature_browse.domain.repositories.paging.CharacterPagingSource
+import ru.kggm.feature_browse.domain.entities.CharacterPagingSource
 
 interface CharacterRepository {
     fun characterPagingSource(): CharacterPagingSource
-    suspend fun getById(id: Int): CharacterEntity
+    suspend fun getById(id: Long): CharacterEntity
 }
